refactor: move functions for sidebar to own module manage_sidebar.js
