+++ nonograms/scripts/game_logic.js
@@ -67,22 +67,6 @@ function setSaveResumeButton(name) {
   }
 }
 
-function toggleSidebar() {
-  const sidebar = document.getElementsByClassName('sidebar');
-  if (!sidebar[0]) {
-    const gameSection = document.getElementsByClassName('game__section');
-    const newSidebar = new Sidebar().createSidebar();
-    gameSection[0].append(newSidebar);
-    listeners.startSidebarListeners();
-  } else {
-    const themeSwitch = document.getElementById('themes__switch');
-    themeSwitch.checked = vars.isDarkTheme;
-    sidebar[0].classList.toggle('sidebar__active');
-  }
-  const isSidebarShown = sidebar[0].classList.contains('sidebar__active');
-  isSidebarShown && refreshBestTimeTable();
-}
-
 function setGameByLevel(event, levelId, table) {
   const eventId = event.target.dataset.id;
   const isResetTime = eventId !== 'save';
@@ -145,25 +129,6 @@ function manageCell(targetCell) {
   const targetCellRow = Number(targetCell.dataset.row);
   const targetCellCol = Number(targetCell.dataset.col);
   setAnswer(sign, targetCellCol, targetCellRow);
-}
-
-function sidebarSettings(event) {
-  const inputType = event.target.type;
-  const inputId = event.target.id;
-  if (inputType === 'checkbox' && inputId === 'themes__switch') {
-    const state = event.target.checked;
-    vars.setThemeState(state);
-    setTheme(state);
-  } else if (inputType === 'checkbox') {
-    const state = event.target.checked;
-    vars.setAudioState(inputId, state);
-  } else if (inputType === 'range') {
-    const value = event.target.value;
-    vars.setAudioVolume(inputId, value);
-  }
-  setTimeout(() => {
-    slConfig.saveAppSettings();
-  }, 1000);
 }
 
 function refreshBestTimeTable() {
